+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * fix(`AbstractState`): Make `$container` protected, so it can be extended by
   child classes
+* build: Create `ApiMapperInterface` and use it for the service ID.
 
 ## 0.5.0